+++ seed/options.js
@@ -1,0 +1,19 @@
+module.exports = {
+    framework: [
+        { name: 'Polymer', value: 'polymer' }
+    ],
+
+    polymer: {
+        compiler: [
+            { name: 'BabelJS', value: 'babeljs' }
+        ],
+
+        css: [
+            { name: 'SaSS', value: 'sass' }
+        ]
+    },
+
+    test: [
+        { name: 'Mocha', value: 'mocha' }
+    ]
+}