+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 - [Spanish](#spanish)
 
 ## English
+### v.0.2.0
+    - Rename project
+    - Updated to generator-speedseed v.0.20.0
+
 ### v.0.1.0
     - Init
 
 ## Spanish
+### v.0.2.0
+    - Actualizado a generator-speedseed v.0.20.0
+    - Renombra el proyecto
+
 ### v.0.1.0
     - Inicia